user 클래스 추가 - 사용자가 구매한 로또 관련 클래스-  및 함수 이동 리팩토링
